+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.28.5
+
+### Fixed
+
+- Handle TextEdit edge case where opcode is checked for 1 past last character. Resolves <https://github.com/pappasam/jedi-language-server/issues/96>
+- No longer return textEdit actions from Jedi that aren't valid Python. Prevents all sorts of wonkyness.
+
 ## 0.28.4
 
 ### Fixed